requirements.txt
django==5.0.2
django-allauth==0.63.3
requests==2.31.0
PyJWT==2.8.0
cryptography==41.0.7
supabase==2.4.0
python-dotenv==1.0.1
